+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * update cache delete and don't use `wp_cache_flush`
+* always show download link in REST endpoint
 
 #### 12.17.3 / 2025-07-31
 * add new `Security` header with value of email or URI